Added login and register pages with components
