+++ src/pages/VerifyEmailPage.jsx
@@ -1,0 +1,10 @@
+import React from 'react';
+import '../assets/css/verify-email.css';
+import VerifyEmail from '../components/Auth/VerifyEmail';
+
+const VerifyEmailPage = () => {
+    return (
+        <VerifyEmail/>
+    );
+};
+export default VerifyEmailPage;
